moved around questions components
