+++ src/Pages/QuestionsPage/ClientDisplayQuestions/ClientDisplayQuestions.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { API, graphqlOperation } from 'aws-amplify';
-import { listQuestions, listQuestionsAnswers } from './graphql/queries';
+import { listQuestions, listQuestionsAnswers } from '../../../graphql/queries';
 
 function ClientDisplayQuestions() {
   const [questions, setQuestions] = useState([]);
